requirements.txt
Flask
python-dotenv
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
openai
requests
beautifulsoup4
